<refactor>[front] Refactor folder structure
- Gruop post related components , server-acion, lib and type in post
  folder
- Minimize folder depth
- Move components in 'modal/auth' to 'authentication' folder
- intergrate _global and common folder
- Delete not used components like <Tap/> and modal related to previous
  login logic
