+++ front/src/app/authentication/sign-in/page.tsx
@@ -1,6 +1,6 @@
 import { Box, Text, Flex } from "@chakra-ui/react";
 import Link from "next/link";
-import Login from "@/components/modal/auth/login";
+import Login from "@/components/authentication/login";
 import { customColors } from "@/utils/chakra/customColors";
 import Brand from "@/components/common/Brand";
