+++ front/src/app/quest/page.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Box } from "@chakra-ui/react";
 import PostWinow from "@/components/posts/card/PostWindow";
-import { POST_TYPE } from "@/type/postType";
+import { POST_TYPE } from "@/components/posts/postType";
 import { customColors } from "@/utils/chakra/customColors";
 export default function QuestPage() {
