+++ front/src/components/authentication/components/common_input.tsx
