+++ front/src/components/authentication/components/toast.tsx
