+++ front/src/components/authentication/login.tsx
@@ -3,8 +3,8 @@ import { Box, Button, Flex, Text } from '@chakra-ui/react';
 import React, { useState } from 'react';
 import { signIn } from '@/server-actions/AuthAction';
 import { useForm } from 'react-hook-form';
-import { CommonInput } from './components/input/common_input';
-import { SignInDTO } from '@/type/server-action/AuthType';
+import { CommonInput } from './components/common_input';
+import { SignInDTO } from '@/type/AuthType';
 import { useRecoilState } from 'recoil';
 import { UserAccountState } from '@/atoms/userAccount';
 import { useRouter } from 'next/navigation';
