+++ front/src/components/comment/server-component/commentArea.tsx
@@ -2,7 +2,7 @@ import InputComment from '../InputComment';
 import { Box, Flex, Spinner, Stack, VStack } from '@chakra-ui/react';
 import { Suspense } from 'react';
 import RootCommentItemList from './rootCommentItemList';
-import { POST_TYPE } from '@/type/postType';
+import { POST_TYPE } from '@/components/posts/postType';
 
 type CommentAreaProps = {
     postType: POST_TYPE;
