+++ front/src/components/common/postSlider.tsx
