--- front/src/components/modal/auth/authInput.tsx
@@ -1,20 +1,0 @@
-import { Flex } from '@chakra-ui/react';
-import React from 'react';
-import { useRecoilValue } from 'recoil';
-import { AuthModalState } from '@/atoms/authModal';
-import Login from './login';
-import SignUp from './signUp';
-
-type AuthInputProps = {};
-
-const AuthInput: React.FC<AuthInputProps> = () => {
-    const modalState = useRecoilValue(AuthModalState);
-
-    return (
-        <Flex direction="column" align="center" width="100%" mt={4} gap="40px">
-            {modalState.view === 'login' && <Login />}
-            {modalState.view === 'signup' && <SignUp />}
-        </Flex>
-    );
-};
-export default AuthInput;
