+++ front/src/components/posts/create/createPostForm.tsx
@@ -6,12 +6,11 @@ import React, { useCallback, useEffect, useState } from 'react';
 import useSelectFile from '@/hooks/useSelectFile';
 import TextInput from './postFormat/textInput';
 import { customColors } from '@/utils/chakra/customColors';
-import Tab, { type TabItem } from './tab';
-import { createPosts } from '@/server-actions/postsActions';
-import { showErrorToast } from '@/components/modal/auth/components/toast/toast';
+import { createPosts } from '@/components/posts/postsActions';
+import { showErrorToast } from '@/components/authentication/components/toast';
 import ImageUploadArea from '@/components/common/ImageUpload';
 import { useUserAccountWithoutSSR } from '@/hooks/userAccount';
-import { NEW_POST_FORMAT, POST_TYPE, NewPostForm, PostContainer, QuestPost, SubmissionPost } from '@/type/postType';
+import { NEW_POST_FORMAT, POST_TYPE, NewPostForm, PostContainer, QuestPost, SubmissionPost } from '@/components/posts/postType';
 import { contentTexts, titleTexts } from '../card/const/rule_card_texts';
 import { useRecoilState } from 'recoil';
 import { UserAccountState } from '@/atoms/userAccount';
