+++ front/src/components/posts/detail/detailPostForm.tsx
@@ -4,12 +4,12 @@ import { customColors } from '@/utils/chakra/customColors';
 import { Box, Button, Divider, Flex, HStack, Icon, Image, Spacer, Tag, Text } from '@chakra-ui/react';
 import DetailPostHeader from './components/DetailPostHeader';
 import ParentPostCard from './components/ParentPostCard';
-import { POST_TYPE, QuestPost, SubmissionPost,PostContainer } from '@/type/postType';
-import { addFavorite, deleteFavorite, evaluateSubmission } from '@/server-actions/postsActions';
+import { POST_TYPE, QuestPost, SubmissionPost,PostContainer } from '@/components/posts/postType';
+import { addFavorite, deleteFavorite, evaluateSubmission } from '@/components/posts/postsActions';
 import { UserAccountState } from '@/atoms/userAccount';
 import { useRecoilState } from 'recoil';
 import { useUserAccountWithSSR, useUserAccountWithoutSSR } from '@/hooks/userAccount';
-import { responseData } from '@/type/server-action/responseType';
+import { responseData } from '@/type/responseType';
 import { useEffect, useState } from 'react';
 import FavoriteButton from '../common/FavoriteButton';
 import { useRouter } from 'next/navigation';
