+++ front/src/components/posts/postType.ts
