+++ front/src/components/posts/postsActions.ts
@@ -9,16 +9,16 @@ import {
     NewPostForm,
     QuestPost,
     SubmissionPost,
-} from '@/type/postType';
+} from '@/components/posts/postType';
 import { revalidateTag } from 'next/cache';
 import { assert, log } from 'console';
 import { validateAuth } from '@/lib/server-only/authLib';
 import { getUserBasicInfo } from '@/lib/server-only/authLib';
 import { validateImage, uploadQuestImage, uploadSubmissionImage } from '@/utils/cloudinary/lib';
 import { cloudinaryTempData } from '@/utils/cloudinary/mockUpData';
-import { responseData } from '@/type/server-action/responseType';
+import { responseData } from '@/type/responseType';
 import { PostApiResponseDto, voteResponseDto } from 'hoodone-shared';
-import { PostCache } from '@/lib/server-only/postLib';
+import { PostCache } from '@/components/posts/postLib';
 import { LoggableResponse } from '@/utils/log/types';
 
 /*
